pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "risk-index"
version = "0.1.0"
description = "Risk-On/Risk-Off Regime Index System"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "yfinance>=0.2.31",
    "fredapi>=0.5",
    "pyarrow>=14.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "plotly>=5.18",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "statsmodels>=0.14",
    "openpyxl>=3.1",
    "kaleido>=0.2",
    "streamlit>=1.28",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "ruff>=0.1"]
email = ["yagmail>=0.15"]

[project.scripts]
risk-update = "scripts.run_update:main"
risk-research = "scripts.run_research:main"
risk-report = "scripts.run_report:main"
risk-dashboard = "scripts.run_dashboard:main"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
